Adding BG stage reason and BG new Student MS score

+++ models/student_new_payments.py
@@ -9,5 +9,5 @@ class StudentNewPaymentModel(db.Model):
 	hubspot_id = db.Column(db.String, nullable=False, unique=True)
 	student_id = db.Column(db.String, nullable=True)
 	type_of_collection = db.Column(db.String, nullable=False)
-	amount = db.Column(db.Float, nullable=False)
+	amount = db.Column(db.Integer, nullable=False)
 	created_at = db.Column(db.DateTime, nullable=False, default=db.func.current_timestamp())
